Brand-new architectural design and ergonomic improve
It is breaking update for WaterUI, introducing a lots of new features and changes.

Now Rust run on main thread, so no thread sync anymore!

Introduce `ConfigurableView`, it allows you to customize the build-in components. And we implements a simple i18n plugin with the new feature.

We also improve ergonomics to reduce boilerplate code. (for example, new utility macros)